+++ packages/client/src/restapi/types/get.ts
@@ -1,34 +1,28 @@
-import { apiRequest, type ApiRequestParams } from '../../API';
-import type { PloneClientConfig } from '../../validation/config';
+import { apiRequest, type ApiRequestParams } from '../../api';
 import { z } from 'zod';
 import type { GetTypeResponse } from '@plone/types';
+import type PloneClient from '../../client';
+import type { RequestResponse } from '../types';
 
 const getTypeSchema = z.object({
   contentPath: z.string(),
 });
 
-export type GetTypeArgs = z.infer<typeof getTypeSchema> & {
-  config: PloneClientConfig;
-};
+export type GetTypeArgs = z.infer<typeof getTypeSchema>;
 
-export const getType = async ({
-  contentPath,
-  config,
-}: GetTypeArgs): Promise<GetTypeResponse> => {
+export async function getType(
+  this: PloneClient,
+  { contentPath }: GetTypeArgs,
+): Promise<RequestResponse<GetTypeResponse>> {
   const validatedArgs = getTypeSchema.parse({
     contentPath,
   });
 
   const options: ApiRequestParams = {
-    config,
+    config: this.config,
     params: {},
   };
   const contentPathPath = `/@types/${validatedArgs.contentPath}`;
 
   return apiRequest('get', contentPathPath, options);
-};
-
-export const getTypeQuery = ({ contentPath, config }: GetTypeArgs) => ({
-  queryKey: [contentPath, 'get', 'types'],
-  queryFn: () => getType({ contentPath, config }),
-});
+}