+++ packages/registry/src/index.ts
@@ -13,6 +13,7 @@ import type {
   UtilitiesConfig,
   ViewsConfig,
   WidgetsConfig,
+  ReactRouterRouteEntry,
 } from '@plone/types';
 
 export type ConfigData = {
@@ -22,6 +23,7 @@ export type ConfigData = {
   widgets: WidgetsConfig | Record<string, never>;
   addonReducers?: AddonReducersConfig;
   addonRoutes?: AddonRoutesConfig;
+  routes?: Array<ReactRouterRouteEntry>;
   slots: SlotsConfig | Record<string, never>;
   components: ComponentsConfig | Record<string, never>;
   utilities: UtilitiesConfig | Record<string, never>;
@@ -124,6 +126,14 @@ class Config {
 
   set addonRoutes(addonRoutes) {
     this._data.addonRoutes = addonRoutes;
+  }
+
+  get routes() {
+    return this._data.routes;
+  }
+
+  set routes(routes) {
+    this._data.routes = routes;
   }
 
   get slots() {
@@ -270,17 +280,18 @@ class Config {
       throw new Error('No component provided');
     }
     if (!predicates) {
-      // Test if there's already one registered, we only support one
       const hasRegisteredNoPredicatesComponent = this._data.slots?.[
         slot
       ]?.data?.[name]?.find(({ predicates }) => !predicates);
+      // If we have one already registered with the same name, and the component
+      // is different, we override it.
+      // During HMR operations when replacing the slot component
+      // errored trying to register it again.
       if (
         hasRegisteredNoPredicatesComponent &&
         component !== hasRegisteredNoPredicatesComponent.component
       ) {
-        throw new Error(
-          `There is already registered a component ${name} for the slot ${slot}. You can only register one slot component with no predicates per slot.`,
-        );
+        hasRegisteredNoPredicatesComponent.component = component;
       }
     }
 
@@ -493,6 +504,12 @@ class Config {
     );
 
     return utilities;
+  }
+
+  registerRoute(options: ReactRouterRouteEntry) {
+    const route = this._data.routes || [];
+    route.push(options);
+    this._data.routes = route;
   }
 }