+++ packages/volto/src/components/manage/Blocks/Block/BlocksForm.jsx
@@ -107,7 +107,11 @@ const BlocksForm = (props) => {
       onFocusNextBlock(block, node, isMultipleSelection);
       e.preventDefault();
     }
-    if (e.key === 'Enter' && !disableEnter) {
+
+    const blockConfig =
+      blocksConfig[properties[getBlocksFieldname(properties)][block]['@type']];
+
+    if (e.key === 'Enter' && !disableEnter && !blockConfig.disableEnter) {
       if (!disableAddBlockOnEnterKey) {
         onSelectBlock(onAddBlock(config.settings.defaultBlockType, index + 1));
       }
@@ -138,7 +142,7 @@ const BlocksForm = (props) => {
   };
 
   const onMutateBlock = (id, value) => {
-    const newFormData = mutateBlock(properties, id, value);
+    const newFormData = mutateBlock(properties, id, value, {}, intl);
     onChangeFormData(newFormData);
   };
 
@@ -149,6 +153,8 @@ const BlocksForm = (props) => {
       value,
       current,
       config.experimental.addBlockButton.enabled ? 1 : 0,
+      {},
+      intl,
     );
 
     const blocksFieldname = getBlocksFieldname(newFormData);
@@ -166,7 +172,7 @@ const BlocksForm = (props) => {
 
   const onAddBlock = (type, index) => {
     if (editable) {
-      const [id, newFormData] = addBlock(properties, type, index);
+      const [id, newFormData] = addBlock(properties, type, index, {}, intl);
       const blocksFieldname = getBlocksFieldname(newFormData);
       const blockData = newFormData[blocksFieldname][id];
       newFormData[blocksFieldname][id] = applyBlockDefaults({
@@ -188,7 +194,7 @@ const BlocksForm = (props) => {
   const onDeleteBlock = (id, selectPrev) => {
     const previous = previousBlockId(properties, id);
 
-    const newFormData = deleteBlock(properties, id);
+    const newFormData = deleteBlock(properties, id, intl);
     onChangeFormData(newFormData);
 
     onSelectBlock(selectPrev ? previous : null);
@@ -260,7 +266,7 @@ const BlocksForm = (props) => {
 
   for (const [n, v] of blockList) {
     if (!v) {
-      const newFormData = deleteBlock(properties, n);
+      const newFormData = deleteBlock(properties, n, intl);
       onChangeFormData(newFormData);
     }
   }