+++ packages/volto/src/components/manage/Blocks/Container/EditBlockWrapper.jsx
@@ -32,21 +32,34 @@ const EditBlockWrapper = (props) => {
     onSelectBlock,
     data,
     index,
+    properties,
   } = blockProps;
 
   function onResetBlock() {
     onChangeBlock(block, { '@type': 'empty' });
   }
 
-  const style = buildStyleObjectFromData(data);
+  const style = buildStyleObjectFromData(
+    data,
+    '',
+    // in a container, we have the parent container data in the properties prop
+    // passing the data of the container too
+    // This is needed in order to calculate properly the styles for the blocks inside the container
+    properties.blocks ? properties : {},
+  );
+
+  // We need to merge the StyleWrapper styles with the draggable props from b-D&D
+  const styleMergedWithDragProps = {
+    ...draginfo.draggableProps,
+    style: { ...style, ...draginfo.draggableProps.style },
+  };
 
   return (
     <div
       ref={draginfo.innerRef}
-      {...draginfo.draggableProps}
+      {...styleMergedWithDragProps}
       {...draginfo.dragHandleProps}
       className={cx(`block-editor-${data['@type']} contained`, { selected })}
-      style={style}
     >
       <div
         role="presentation"
@@ -61,6 +74,7 @@ const EditBlockWrapper = (props) => {
             aria-label={intl.formatMessage(messages.reset, {
               index,
             })}
+            type="button"
             basic
             icon
             onClick={(e) => onResetBlock(block, {})}
@@ -70,6 +84,7 @@ const EditBlockWrapper = (props) => {
           </Button>
         ) : (
           <Button
+            type="button"
             basic
             icon
             className="remove-block-button"