+++ packages/volto/src/components/manage/Blocks/Search/hocs/withSearch.jsx
@@ -18,6 +18,7 @@ const SEARCH_ENDPOINT_FIELDS = [
   'limit',
   'sort_on',
   'sort_order',
+  'depth',
 ];
 
 const PAQO = 'plone.app.querystring.operation';
@@ -73,6 +74,7 @@ function getInitialState(
     sort_order: sortOrderParam || data.query?.sort_order,
     b_size: data.query?.b_size,
     limit: data.query?.limit,
+    depth: data.query?.depth,
     block: id,
   };
 }
@@ -130,6 +132,7 @@ function normalizeState({
     sort_order: sortOrder || query.sort_order,
     b_size: query.b_size,
     limit: query.limit,
+    depth: query.depth,
     block: id,
   };