+++ packages/volto/src/components/manage/Contents/Contents.jsx
@@ -456,6 +456,7 @@ class Contents extends Component {
       this.setState(
         {
           currentPage: 0,
+          selected: [],
         },
         () =>
           this.setState({ filter: '' }, () =>
@@ -652,6 +653,7 @@ class Contents extends Component {
     this.setState(
       {
         currentPage: value,
+        selected: [],
       },
       () => this.fetchContents(),
     );
@@ -669,6 +671,7 @@ class Contents extends Component {
       {
         pageSize: value,
         currentPage: 0,
+        selected: [],
       },
       () => this.fetchContents(),
     );
@@ -728,6 +731,7 @@ class Contents extends Component {
     this.setState({
       sort_on: values[0],
       sort_order: values[1],
+      selected: [],
     });
     this.props.sortContent(
       getBaseUrl(this.props.pathname),
@@ -755,6 +759,7 @@ class Contents extends Component {
         this.setState(
           {
             currentPage: 0,
+            selected: [],
           },
           () => this.fetchContents(),
         );
@@ -780,6 +785,7 @@ class Contents extends Component {
         this.setState(
           {
             currentPage: 0,
+            selected: [],
           },
           () => this.fetchContents(),
         );