+++ packages/volto/src/components/manage/Controlpanels/Relations/Relations.jsx
@@ -47,53 +47,55 @@ const RelationsControlPanel = () => {
 
   return (
     <>
-      <div className="relations-control-panel">
-        <Helmet title={intl.formatMessage(messages.relations)} />
-        {can_edit ? (
-          <Segment.Group raised>
-            <Segment className="primary">
-              {brokenRelations && Object.keys(brokenRelations).length > 0 ? (
-                <React.Fragment>
-                  <Message warning>
-                    <FormattedMessage
-                      id="Some relations are broken. Please fix."
-                      defaultMessage="Some relations are broken. Please fix."
-                    />
-                  </Message>
-                  <Divider hidden />
-                </React.Fragment>
-              ) : null}
-              <h1>
+      <div className="ui container">
+        <div className="relations-control-panel">
+          <Helmet title={intl.formatMessage(messages.relations)} />
+          {can_edit ? (
+            <Segment.Group raised>
+              <Segment className="primary">
+                {brokenRelations && Object.keys(brokenRelations).length > 0 ? (
+                  <React.Fragment>
+                    <Message warning>
+                      <FormattedMessage
+                        id="Some relations are broken. Please fix."
+                        defaultMessage="Some relations are broken. Please fix."
+                      />
+                    </Message>
+                    <Divider hidden />
+                  </React.Fragment>
+                ) : null}
+                <h1>
+                  <FormattedMessage id="Relations" defaultMessage="Relations" />
+                </h1>
+                {relations_stats?.error ? (
+                  <React.Fragment>
+                    <Divider hidden />
+                    <Message warning>
+                      <FormattedMessage
+                        id="Please upgrade to plone.restapi >= 8.39.0."
+                        defaultMessage="Please upgrade to plone.restapi >= 8.39.0."
+                      />
+                    </Message>
+                  </React.Fragment>
+                ) : null}
+              </Segment>
+              <Segment>
+                <RelationsMatrix />
+              </Segment>
+            </Segment.Group>
+          ) : (
+            <Segment.Group>
+              <Segment>
                 <FormattedMessage id="Relations" defaultMessage="Relations" />
-              </h1>
-              {relations_stats?.error ? (
-                <React.Fragment>
-                  <Divider hidden />
-                  <Message warning>
-                    <FormattedMessage
-                      id="Please upgrade to plone.restapi >= 8.39.0."
-                      defaultMessage="Please upgrade to plone.restapi >= 8.39.0."
-                    />
-                  </Message>
-                </React.Fragment>
-              ) : null}
-            </Segment>
-            <Segment>
-              <RelationsMatrix />
-            </Segment>
-          </Segment.Group>
-        ) : (
-          <Segment.Group>
-            <Segment>
-              <FormattedMessage id="Relations" defaultMessage="Relations" />
-              <Divider hidden />
-              <FormattedMessage
-                id="You have not the required permission for this control panel."
-                defaultMessage="You have not the required permission for this control panel."
-              />
-            </Segment>
-          </Segment.Group>
-        )}
+                <Divider hidden />
+                <FormattedMessage
+                  id="You have not the required permission for this control panel."
+                  defaultMessage="You have not the required permission for this control panel."
+                />
+              </Segment>
+            </Segment.Group>
+          )}
+        </div>
       </div>
 
       {isClient &&