+++ packages/volto/src/config/index.js
@@ -113,7 +113,7 @@ let config = {
     defaultPageSize: 25,
     isMultilingual: false,
     supportedLanguages: ['en'],
-    defaultLanguage: 'en',
+    defaultLanguage: process.env.SITE_DEFAULT_LANGUAGE || 'en',
     navDepth: 1,
     expressMiddleware: serverConfig.expressMiddleware, // BBB
     defaultBlockType: 'slate',
@@ -141,6 +141,7 @@ let config = {
     serverConfig,
     storeExtenders: [],
     showTags: true,
+    showRelatedItems: false,
     controlpanels: [],
     controlPanelsIcons,
     filterControlPanels,
@@ -162,7 +163,6 @@ let config = {
     showSelfRegistration: false,
     contentMetadataTagsImageField: 'image',
     contentPropertiesSchemaEnhancer: null,
-    hasWorkingCopySupport: false,
     maxUndoLevels: 200, // undo history size for the main form
     addonsInfo: addonsInfo,
     workflowMapping,
@@ -231,6 +231,16 @@ Object.entries(slots).forEach(([slotName, components]) => {
     });
   });
 });
+
+// Make sure that process.env.SITE_DEFAULT_LANGUAGE is set in availableLanguages
+if (
+  process.env.SITE_DEFAULT_LANGUAGE &&
+  !config.settings.supportedLanguages.includes(
+    process.env.SITE_DEFAULT_LANGUAGE,
+  )
+) {
+  config.settings.supportedLanguages.push(process.env.SITE_DEFAULT_LANGUAGE);
+}
 
 registerValidators(ConfigRegistry);
 installDefaultComponents(ConfigRegistry);