+++ packages/volto/src/config/validation.ts
@@ -15,6 +15,7 @@ import {
   startEventDateRangeValidator,
   endEventDateRangeValidator,
   patternValidator,
+  defaultLanguageControlPanelValidator,
 } from '@plone/volto/helpers/FormValidation/validators';
 
 const registerValidators = (config: ConfigType) => {
@@ -149,6 +150,13 @@ const registerValidators = (config: ConfigType) => {
     type: 'validator',
     dependencies: { behaviorName: 'plone.eventbasic', fieldName: 'end' },
     method: endEventDateRangeValidator,
+  });
+
+  config.registerUtility({
+    name: 'default_language',
+    type: 'validator',
+    dependencies: { format: 'default_language' },
+    method: defaultLanguageControlPanelValidator,
   });
 };