+++ packages/volto/src/express-middleware/robotstxt.js
@@ -4,13 +4,12 @@ import { generateRobots } from '@plone/volto/helpers/Robots/Robots';
 /*
 robots.txt - priority order:
 
-1) robots.txt in /public folder
-2) VOLTO_ROBOTSTXT var in .env
-3) default: plone robots.txt
+1) VOLTO_ROBOTSTXT var in .env
+2) robots.txt setting in the site control panel
 
 */
 
-const ploneRobots = function (req, res, next) {
+const siteRobots = function (req, res, next) {
   generateRobots(req).then((robots) => {
     res.set('Content-Type', 'text/plain');
     res.send(robots);
@@ -27,7 +26,7 @@ export default function robotstxtMiddleware() {
   if (process.env.VOLTO_ROBOTSTXT) {
     middleware.all('**/robots.txt', envRobots);
   } else {
-    middleware.all('**/robots.txt', ploneRobots);
+    middleware.all('**/robots.txt', siteRobots);
   }
   middleware.id = 'robots.txt';
   return middleware;