+++ packages/client/CHANGELOG.md
@@ -8,6 +8,28 @@
 
 <!-- towncrier release notes start -->
 
+## 1.0.0-alpha.23 (2025-02-08)
+
+### Internal
+
+- Update Vite version. @sneridagh [#6640](https://github.com/plone/volto/pull/6640)
+- Update internal `peerDependencies` to include React 19.
+  Update TS version. @sneridagh [#6641](https://github.com/plone/volto/pull/6641)
+- Test with Plone 6.1.0rc1. @sneridagh [#6682](https://github.com/plone/volto/pull/6682)
+- Remove hard dependencies on `react`, use `peerDependencies` instead. @sneridagh [#6728](https://github.com/plone/volto/pull/6728)
+
+## 1.0.0-alpha.22 (2025-01-24)
+
+### Documentation
+
+- Fix typo in README. @sneridagh [#6599](https://github.com/plone/volto/pull/6599)
+
+## 1.0.0-alpha.21 (2025-01-15)
+
+### Feature
+
+- Import `getContent` bare fetcher. @sneridagh [#6594](https://github.com/plone/volto/pull/6594)
+
 ## 1.0.0-alpha.20 (2024-11-05)
 
 ### Internal