+++ packages/helpers/CHANGELOG.md
@@ -8,6 +8,19 @@
 
 <!-- towncrier release notes start -->
 
+## 1.0.2 (2025-02-08)
+
+### Internal
+
+- Update internal `peerDependencies` to include React 19.
+  Update TS version. @sneridagh [#6641](https://github.com/plone/volto/issues/6641)
+
+## 1.0.1 (2025-01-24)
+
+### Internal
+
+- Centralize `tsconfig`. @sneridagh [#6536](https://github.com/plone/volto/issues/6536)
+
 ## 1.0.0 (2024-12-12)
 
 ### Internal